pyproject.toml
[tool.poetry]
name = "otp"
version = "0.1.0"
description = ""
authors = ["Reganto <tell.reganto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.14"
djangorestframework = "^3.13.1"
django-ptpython = "^1.0.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
